04_entrevistas/01_entrevistas_cruas/stefani/requirements.txt
PyPDF2
pytesseract
Pillow
